MedReminder/requirements.txt
google-genai
pydantic>=2.0.0
boto3>=1.28.0
